+++ PHC-Android/PHC/src/main/res/layout/fragment_checkout_form.xml
@@ -1,0 +1,161 @@
+
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/checkout_form"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center_horizontal"
+    android:layout_marginTop="@dimen/top_bottom_margins"
+    android:layout_marginBottom="@dimen/top_bottom_margins">
+    <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        android:id="@+id/checkout_layout"
+        android:orientation="vertical"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:gravity="center"
+            android:textSize="20sp"
+            android:text="@string/checkout_experience_title"
+            />
+        <RadioGroup
+            android:id="@+id/checkout_radiogroup"
+            android:layout_width="fill_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+            <RadioButton
+                android:id="@+id/checkout_radio_0"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_0"
+                android:orientation="horizontal"
+                android:layout_weight="1"
+                android:padding="10dp"
+                />
+            <RadioButton
+                android:id="@+id/checkout_radio_1"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_1"
+                android:layout_weight="1"
+                android:orientation="horizontal"
+                android:padding="10dp"
+                />
+            <RadioButton
+                android:id="@+id/checkout_radio_2"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_2"
+                android:layout_weight="1"
+                android:orientation="horizontal"
+                android:padding="10dp"
+                />
+            <RadioButton
+                android:id="@+id/checkout_radio_3"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_3"
+                android:layout_weight="1"
+                android:orientation="horizontal"
+                android:padding="10dp"
+                />
+            <RadioButton
+                android:id="@+id/checkout_radio_4"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_4"
+                android:layout_weight="1"
+                android:orientation="horizontal"
+                android:padding="10dp"
+                />
+            <RadioButton
+                android:id="@+id/checkout_radio_5"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/checkout_experience_5"
+                android:layout_weight="1"
+                android:orientation="horizontal"
+                android:padding="10dp"
+                />
+        </RadioGroup>
+        <!--
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:gravity="center"
+            android:id="@+id/exit_services_title"
+            android:textSize="20sp"
+            android:text="@string/checkout_services_title"
+            />
+
+        <CheckBox
+            android:id="@+id/exit_service_holder"
+            android:layout_width="match_parent"
+            android:layout_height="20dp"
+            android:layout_margin="10dp"
+            android:text="@string/checkout_service_blank"
+            />
+            -->
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:gravity="center"
+            android:id="@+id/experience_text"
+            android:textSize="20sp"
+            android:text="@string/checkout_comment_title"
+            />
+
+        <EditText
+            android:id="@+id/checkout_comment"
+            android:layout_width="match_parent"
+            android:layout_height="100dp"
+            android:layout_margin="10dp"
+            android:inputType="textCapWords|textMultiLine"
+            android:textSize="20sp"
+            android:maxLines="10"/>
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="10dp"
+            android:gravity="center"
+            android:id="@+id/enter_code"
+            android:textSize="20sp"
+            android:text="@string/enter_code_prompt"
+            />
+
+        <EditText
+            android:id="@+id/text_code"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:inputType="phone"
+            android:layout_gravity="center"
+            android:textSize="20sp"
+            android:maxLength="10"
+            android:singleLine="true"/>
+        <Button
+            android:layout_gravity="center"
+            android:id="@+id/button_submit"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/button_submit"
+            android:textSize="30sp"/>
+
+
+        <!--
+        <Button
+            android:id="@+id/exit_qr_scanner"
+            android:text="@string/exit_qr_scanner"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            />
+            -->
+
+
+    </LinearLayout>
+</ScrollView>
