+++ PHC-Android/PHC/src/main/res/layout/fragment_checkout_success.xml
@@ -6,7 +6,9 @@
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
     android:layout_gravity="center_horizontal"
-    android:layout_marginTop="@dimen/selection_top_margin">
+    android:layout_marginTop="@dimen/selection_top_margin"
+    android:id="@+id/checkout_success"
+    >
 
     <TextView
         android:id="@+id/text_success"
