+++ PHC-Android/PHC/src/main/res/values/strings.xml
@@ -170,4 +170,6 @@
     <string name="checkout_experience_3">3</string>
     <string name="checkout_experience_4">4</string>
     <string name="checkout_experience_5">5</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
 </resources>
